Create manual-autoplay.js to extract manualAutoplay function from player.js
Issue #2
